pyproject.toml
[project]
name = "p10-oc-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Charles DZADU",email = "dzaducharles@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.3.0,<6.0.0)",
    "django-cors-headers (>=4.3.0,<5.0.0)",
    "python-dateutil (>=2.8.2,<3.0.0)",
    "drf-spectacular (>=0.27.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
